backend/gpt/requirements.txt
pinecone
openai
langchain
python-dotenv
